+++ mm2src/mm2_main/src/mm2_tests/structs.rs
@@ -619,6 +619,16 @@ pub enum EnableCoinBalance {
 pub struct ZcoinActivationResult {
     pub current_block: u64,
     pub wallet_balance: EnableCoinBalance,
+    #[serde(default)]
+    pub first_sync_block: Option<FirstSyncBlock>,
+}
+
+#[derive(Debug, Deserialize)]
+#[serde(deny_unknown_fields)]
+pub struct FirstSyncBlock {
+    pub requested: u64,
+    pub is_pre_sapling: bool,
+    pub actual: u64,
 }
 
 #[derive(Debug, Deserialize)]